@@ -4,28 +4,28 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
+  sdk: ">=2.19.0 <3.0.0"
 dependencies:
   flutter:
     sdk: flutter
-  better_player: ^0.0.81
+  better_player: ^0.0.83
   shelf_static: ^1.1.1
-  shelf_router: ^1.0.0
-  url_launcher: ^6.0.4
-  get: ^4.1.4
-  dio: ^4.0.0
+  shelf_router: ^1.1.3
+  url_launcher: ^6.1.8
+  get: ^4.6.5
+  dio: ^4.0.6
   global_repository:
     git: https://github.com/nightmare-space/global_repository
-  path: ^1.8.0
+  path: ^1.8.2
   open_file: ^3.2.1
-  cupertino_icons: ^1.0.2
-  signale: ^0.0.6
-  flutter_svg: ^0.22.0
+  cupertino_icons: ^1.0.5
+  signale: ^0.0.8
+  flutter_svg: ^1.1.6
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
+  flutter_lints: ^2.0.1
 
 flutter:
   uses-material-design: true
