@@ -1,6 +1,6 @@
 class Config {
   static String packageName = 'com.example.file_manager_view';
   static String flutterPackage = '';
-  static String package;
+  static String ?package;
   static String packagePrefix = 'packages/file_manager_view/';
 }
