@@ -1,4 +1,4 @@
 // ignore: avoid_web_libraries_in_flutter
 import 'dart:html';
 
-String url = window.document.baseUri;
+String ?url = window.document.baseUri;
